Helper function to convert the data-object into a tibble

as.data.frame() is a function that converts the tidyproteomics data object into a tibble. This tibble is in the long-format, such that a there is a single observation per line.

Usage

# S3 method for tidyproteomics
as.data.frame(data, shape = c("long", "wide"), values = NULL)

Arguments

data

tidyproteomics data object

shape

the orientation of the quantitative data as either a single measure per row (long), or as multiple measures per protein/peptide (wide).

values

indicates the selected normalization to output. The default is that selected at the time of normalization.

Value

a tibble

Examples

library(dplyr, warn.conflicts = FALSE)
library(tidyproteomics)

# convert the data-object to a data.frame
hela_proteins %>% as.data.frame() %>% as_tibble()
#> # A tibble: 42,330 × 20
#>    sample_id sample_file sample  replicate protein    abundance_raw description 
#>    <chr>     <chr>       <chr>   <chr>     <chr>              <dbl> <chr>       
#>  1 e9b20ea7  f1          control 1         Q15149       1011259992. Plectin OS=…
#>  2 e9b20ea7  f1          control 1         Q09666        659299359. Neuroblast …
#>  3 e9b20ea7  f1          control 1         H0YDN1           239911. Plectin OS=…
#>  4 e9b20ea7  f1          control 1         Q14204        385570969  Cytoplasmic…
#>  5 e9b20ea7  f1          control 1         P49327       1441909420. Fatty acid …
#>  6 e9b20ea7  f1          control 1         P78527        622951615. DNA-depende…
#>  7 e9b20ea7  f1          control 1         P21333        604798069. Filamin-A O…
#>  8 e9b20ea7  f1          control 1         P35579       1070673931. Myosin-9 OS…
#>  9 e9b20ea7  f1          control 1         O75369        325748424. Filamin-B O…
#> 10 e9b20ea7  f1          control 1         A0A0D9SGF6    196230866. Spectrin al…
#> # ℹ 42,320 more rows
#> # ℹ 13 more variables: gene_id_entrez <chr>, gene_id_ensemble <chr>,
#> #   gene_name <chr>, biological_process <chr>, cellular_component <chr>,
#> #   molecular_function <chr>, wiki_pathway <chr>, reactome_pathway <chr>,
#> #   protein_group <chr>, match_between_runs <lgl>, num_peptides <dbl>,
#> #   num_unique_peptides <dbl>, num_psms <dbl>

# select the wide format
hela_proteins %>% as.data.frame(shape = 'wide') %>% as_tibble()
#> # A tibble: 9,434 × 21
#>    protein    description              gene_id_entrez gene_id_ensemble gene_name
#>    <chr>      <chr>                    <chr>          <chr>            <chr>    
#>  1 Q15149     Plectin OS=Homo sapiens… 5339           ENSG00000178209  PLEC     
#>  2 Q09666     Neuroblast differentiat… 79026          ENSG00000124942  AHNAK    
#>  3 H0YDN1     Plectin OS=Homo sapiens… NA             ENSG00000178209  PLEC     
#>  4 Q14204     Cytoplasmic dynein 1 he… 1778           ENSG00000197102  DYNC1H1  
#>  5 P49327     Fatty acid synthase OS=… 2194           ENSG00000169710  FASN     
#>  6 P78527     DNA-dependent protein k… 5591           ENSG00000253729  PRKDC    
#>  7 P21333     Filamin-A OS=Homo sapie… 2316           ENSG00000196924  FLNA     
#>  8 P35579     Myosin-9 OS=Homo sapien… 4627           ENSG00000100345  MYH9     
#>  9 O75369     Filamin-B OS=Homo sapie… 2317           ENSG00000136068  FLNB     
#> 10 A0A0D9SGF6 Spectrin alpha chain, n… NA             ENSG00000197694  SPTAN1   
#> # ℹ 9,424 more rows
#> # ℹ 16 more variables: biological_process <chr>, cellular_component <chr>,
#> #   molecular_function <chr>, wiki_pathway <chr>, reactome_pathway <chr>,
#> #   protein_group <chr>, match_between_runs <lgl>, num_peptides <dbl>,
#> #   num_unique_peptides <dbl>, num_psms <dbl>, control_1_raw <dbl>,
#> #   control_2_raw <dbl>, control_3_raw <dbl>, knockdown_1_raw <dbl>,
#> #   knockdown_2_raw <dbl>, knockdown_3_raw <dbl>