Compute protein enrichment

enrichment() is an analysis function that computes the protein summary statistics for a given tidyproteomics data object.

Usage

enrichment(
  data = NULL,
  ...,
  .term = NULL,
  .method = c("gsea", "wilcoxon"),
  .score_type = c("std", "pos", "neg"),
  .cpu_cores = 1
)

Arguments

data

tidyproteomics data object

...

two sample comparison e.g. experimental/control

.term

a character string referencing ".term" in the annotations table

.method

a character string

.score_type

a character string. From the fgsea manual: "This parameter defines the GSEA score type. Possible options are ("std", "pos", "neg"). By default ("std") the enrichment score is computed as in the original GSEA. The "pos" and "neg" score types are intended to be used for one-tailed tests (i.e. when one is interested only in positive ("pos") or negateive ("neg") enrichment)."

.cpu_cores

the number of threads used to speed the calculation

Value

a tibble

Examples

library(dplyr, warn.conflicts = FALSE)
library(tidyproteomics)

# using the default GSEA method
hela_proteins %>%
   expression(knockdown/control) %>%
   enrichment(knockdown/control, .term = "biological_process") %>%
   export_analysis(knockdown/control, .analysis = "enrichment", .term = "biological_process")
#> Warning: All values in the stats vector are greater than zero and scoreType is "std", maybe you should switch to scoreType = "pos".
#> Warning: All values in the stats vector are greater than zero and scoreType is "std", maybe you should switch to scoreType = "pos".
#> Warning: All values in the stats vector are greater than zero and scoreType is "std", maybe you should switch to scoreType = "pos".
#> Warning: All values in the stats vector are greater than zero and scoreType is "std", maybe you should switch to scoreType = "pos".
#> Warning: All values in the stats vector are greater than zero and scoreType is "std", maybe you should switch to scoreType = "pos".
#> Warning: All values in the stats vector are greater than zero and scoreType is "std", maybe you should switch to scoreType = "pos".
#> Warning: All values in the stats vector are greater than zero and scoreType is "std", maybe you should switch to scoreType = "pos".
#> Warning: All values in the stats vector are greater than zero and scoreType is "std", maybe you should switch to scoreType = "pos".
#> Warning: All values in the stats vector are greater than zero and scoreType is "std", maybe you should switch to scoreType = "pos".
#> Warning: All values in the stats vector are greater than zero and scoreType is "std", maybe you should switch to scoreType = "pos".
#> Warning: All values in the stats vector are greater than zero and scoreType is "std", maybe you should switch to scoreType = "pos".
#> Warning: All values in the stats vector are greater than zero and scoreType is "std", maybe you should switch to scoreType = "pos".
#> Warning: All values in the stats vector are greater than zero and scoreType is "std", maybe you should switch to scoreType = "pos".
#> Warning: All values in the stats vector are greater than zero and scoreType is "std", maybe you should switch to scoreType = "pos".
#> # A tibble: 13 × 7
#>    annotation p_value adj_p_value enrichment enrichment_normalized log2err  size
#>    <chr>        <dbl>       <dbl>      <dbl>                 <dbl>   <dbl> <int>
#>  1 cell prol… 0.00327      0.0426      0.333                 1.24  0.432     279
#>  2 cell diff… 0.108        1           0.286                 1.09  0.131     520
#>  3 cellular … 0.156        1           0.274                 1.06  0.106     954
#>  4 coagulati… 0.579        1           0.256                 0.988 0.0389    889
#>  5 cell comm… 0.658        1           0.263                 0.957 0.0329    145
#>  6 cell death 0.724        1           0.257                 0.937 0.0282    160
#>  7 cellular … 0.783        1           0.248                 0.965 0.0240   2761
#>  8 cell orga… 0.815        1           0.246                 0.954 0.0217   1273
#>  9 defense r… 0.926        1           0.238                 0.915 0.0129    789
#> 10 metabolic… 0.975        1           0.232                 0.905 0.00730  2965
#> 11 developme… 0.990        1           0.221                 0.851 0.00458   801
#> 12 cell grow… 1            1           0.221                 0.857 0        1671
#> 13 conjugati… 1            1           0.205                 0.795 0        1150

# using a Wilcoxon Rank Sum method
hela_proteins %>%
   expression(knockdown/control) %>%
   enrichment(knockdown/control, .term = "biological_process", .method = "wilcoxon") %>%
   export_analysis(knockdown/control, .analysis = "enrichment", .term = "biological_process")
#> ℹ annotation other had issues, not reported
#> # A tibble: 13 × 5
#>    annotation                          p_value adj_p_value enrichment  size
#>    <chr>                                 <dbl>       <dbl>      <dbl> <int>
#>  1 conjugation                      0.00000342   0.0000445      1.12   1150
#>  2 cell proliferation               0.000389     0.00467        0.801   279
#>  3 development                      0.00272      0.0299         1.11    801
#>  4 cellular component movement      0.00592      0.0592         0.926   954
#>  5 cell organization and biogenesis 0.00867      0.0780         1.05   1273
#>  6 defense response                 0.0206       0.165          1.08    789
#>  7 coagulation                      0.0769       0.538          1.09    889
#>  8 cell growth                      0.0992       0.595          1.06   1671
#>  9 metabolic process                0.116        0.595          1.03   2965
#> 10 cell differentiation             0.161        0.645          0.956   520
#> 11 cell communication               0.470        1              1.07    145
#> 12 cell death                       0.792        1              1.05    160
#> 13 cellular homeostasis             0.796        1              1.00   2761