Export the quantitative data from an tidyproteomics data-object

export() returns the main quantitative data object as a tibble with identifier as the designation for the measured observation.

Usage

export_analysis(
  data = NULL,
  ...,
  .analysis = NULL,
  .term = NULL,
  .append = NULL,
  .file_name = NULL
)

Arguments

data

tidyproteomics data object

...

two sample comparison e.g. experimental/control

.analysis

a character string for the specific analysis to export

.term

a character string of the term from an enrichment analysis

.append

a character string of the term to append to the output

.file_name

a character string for file to write to, format implied from string ('.rds', '.xlsx', '.csv', '.tsv')

Value

a tibble

Examples

library(dplyr, warn.conflicts = FALSE)
library(tidyproteomics)
hela_proteins %>%
   expression(knockdown/control) %>%
   export_analysis(knockdown/control,
                   .analysis = "expression")
#> # A tibble: 4,829 × 18
#>    protein imputed     n avera…¹ propo…² foldc…³ log2_…⁴ p_value adj_p…⁵ descr…⁶
#>    <chr>     <dbl> <int>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl> <chr>  
#>  1 Q9NXV2    0         6  3.10e6 1.20e-5   19.6     4.30 0.00417       1 BTB/PO…
#>  2 Q8TBM8    0.333     6  3.37e6 1.31e-5    8.72    3.12 0.00437       1 DnaJ h…
#>  3 Q5JUW8    0         6  1.75e6 6.76e-6    7.98    3.00 0.0232        1 Discs,…
#>  4 O95833    0.167     6  3.42e6 1.33e-5    7.13    2.83 0.0298        1 Chlori…
#>  5 Q9Y385    0         6  1.42e6 5.49e-6    7.10    2.83 0.0126        1 Ubiqui…
#>  6 Q14534    0.333     6  4.56e6 1.77e-5    7.02    2.81 0.0154        1 Squale…
#>  7 O00214    0.333     6  1.33e6 5.16e-6    6.49    2.70 0.00818       1 Galect…
#>  8 Q9Y4C2    0         6  7.29e6 2.83e-5    5.78    2.53 0.0152        1 TRPM8 …
#>  9 Q8WU67    0.333     6  9.60e5 3.72e-6    5.77    2.53 0.0763        1 Phosph…
#> 10 Q9H6H4    0         6  2.33e6 9.02e-6    5.72    2.52 0.255         1 Recept…
#> # … with 4,819 more rows, 8 more variables: gene_id_entrez <chr>,
#> #   gene_id_ensemble <chr>, gene_name <chr>, biological_process <chr>,
#> #   cellular_component <chr>, molecular_function <chr>, wiki_pathway <chr>,
#> #   reactome_pathway <chr>, and abbreviated variable names ¹​average_expression,
#> #   ²​proportional_expression, ³​foldchange, ⁴​log2_foldchange, ⁵​adj_p_value,
#> #   ⁶​description

hela_proteins %>%
   export_analysis(.analysis = "counts")
#> # A tibble: 12 × 9
#>    sample_id sample    replicate protei…¹ prote…² pepti…³ pepti…⁴ quant…⁵ is_mbr
#>    <chr>     <chr>     <chr>        <int>   <int>   <dbl>   <dbl>   <dbl> <lgl> 
#>  1 e9b20ea7  control   1             5702    5702   62778   56459   0.939 FALSE 
#>  2 ea36dac9  knockdown 2             5786    5786   63361   56797   0.939 FALSE 
#>  3 ebf4b0fe  knockdown 1             5825    5825   63349   56846   0.937 FALSE 
#>  4 ecfd1822  knockdown 3             5817    5817   63247   56812   0.936 FALSE 
#>  5 eebba67b  control   3             5729    5729   62926   56523   0.942 FALSE 
#>  6 ef79cc4c  control   2             5764    5764   62907   56550   0.939 FALSE 
#>  7 e9b20ea7  control   1             7055    7055   66329   58706   0.907 TRUE  
#>  8 ea36dac9  knockdown 2             7055    7055   66329   58706   0.91  TRUE  
#>  9 ebf4b0fe  knockdown 1             7055    7055   66329   58706   0.909 TRUE  
#> 10 ecfd1822  knockdown 3             7055    7055   66329   58706   0.907 TRUE  
#> 11 eebba67b  control   3             7055    7055   66329   58706   0.91  TRUE  
#> 12 ef79cc4c  control   2             7055    7055   66329   58706   0.907 TRUE  
#> # … with abbreviated variable names ¹​protein_groups, ²​proteins, ³​peptides,
#> #   ⁴​peptides_unique, ⁵​quantifiable