Summarize the data

expression() is an analysis function that computes the protein summary statistics for a given tidyproteomics data object.

Usage

expression(data = NULL, ..., .method = stats::t.test, .p.adjust = "BH")

Arguments

data

tidyproteomics data object

...

two sample comparison e.g. experimental/control

.method

a two-distribution test function returning a p_value for the null hypothesis. Example functions include t.test, wilcox.test, stats::ks.test, additionally, the string "limma" can be used to select from the limma package to compute an empirical Bayesian estimation which performs better with non-linear distributions and uneven replicate balance between samples.

.p.adjust

a stats::p.adjust string for multiple test correction, default is 'BH' (Benjamini & Hochberg, 1995)

Value

a tibble

Examples

library(dplyr, warn.conflicts = FALSE)
library(tidyproteomics)

# simple t.test expression analysis
hela_proteins %>%
   expression(knockdown/control) %>%
   export_analysis(knockdown/control, .analysis = "expression")
#> # A tibble: 4,829 × 18
#>    protein imputed     n average_expression proportional_expression foldchange
#>    <chr>     <dbl> <int>              <dbl>                   <dbl>      <dbl>
#>  1 Q9NXV2    0         6           3103100.              0.0000120       19.6 
#>  2 Q8TBM8    0.333     6           3374207.              0.0000131        8.72
#>  3 Q5JUW8    0         6           1745373.              0.00000676       7.98
#>  4 O95833    0.167     6           3424853.              0.0000133        7.13
#>  5 Q9Y385    0         6           1415381.              0.00000549       7.10
#>  6 Q14534    0.333     6           4560342.              0.0000177        7.02
#>  7 O00214    0.333     6           1332337.              0.00000516       6.49
#>  8 Q9Y4C2    0         6           7289389.              0.0000283        5.78
#>  9 Q8WU67    0.333     6            960150.              0.00000372       5.77
#> 10 Q9H6H4    0         6           2328479.              0.00000902       5.72
#> # ℹ 4,819 more rows
#> # ℹ 12 more variables: log2_foldchange <dbl>, p_value <dbl>, adj_p_value <dbl>,
#> #   description <chr>, gene_id_entrez <chr>, gene_id_ensemble <chr>,
#> #   gene_name <chr>, biological_process <chr>, cellular_component <chr>,
#> #   molecular_function <chr>, wiki_pathway <chr>, reactome_pathway <chr>

# a wilcox.test expression analysis
hela_proteins %>%
   expression(knockdown/control, .method = stats::wilcox.test) %>%
   export_analysis(knockdown/control, .analysis = "expression")
#> # A tibble: 4,829 × 18
#>    protein imputed     n average_expression proportional_expression foldchange
#>    <chr>     <dbl> <int>              <dbl>                   <dbl>      <dbl>
#>  1 Q9NXV2    0         6           3103100.              0.0000120       19.6 
#>  2 Q8TBM8    0.333     6           3374207.              0.0000131        8.72
#>  3 Q5JUW8    0         6           1745373.              0.00000676       7.98
#>  4 O95833    0.167     6           3424853.              0.0000133        7.13
#>  5 Q9Y385    0         6           1415381.              0.00000549       7.10
#>  6 Q14534    0.333     6           4560342.              0.0000177        7.02
#>  7 O00214    0.333     6           1332337.              0.00000516       6.49
#>  8 Q9Y4C2    0         6           7289389.              0.0000283        5.78
#>  9 Q8WU67    0.333     6            960150.              0.00000372       5.77
#> 10 Q9H6H4    0         6           2328479.              0.00000902       5.72
#> # ℹ 4,819 more rows
#> # ℹ 12 more variables: log2_foldchange <dbl>, p_value <dbl>, adj_p_value <dbl>,
#> #   description <chr>, gene_id_entrez <chr>, gene_id_ensemble <chr>,
#> #   gene_name <chr>, biological_process <chr>, cellular_component <chr>,
#> #   molecular_function <chr>, wiki_pathway <chr>, reactome_pathway <chr>

# a one-tailed wilcox.test expression analysis
wilcoxon_less <- function(x, y) {
   stats::wilcox.test(x, y, alternative = "less")
}
hela_proteins <- hela_proteins %>%
   expression(knockdown/control, .method = stats::wilcox.test)

hela_proteins %>% export_analysis(knockdown/control, .analysis = "expression")
#> # A tibble: 4,829 × 18
#>    protein imputed     n average_expression proportional_expression foldchange
#>    <chr>     <dbl> <int>              <dbl>                   <dbl>      <dbl>
#>  1 Q9NXV2    0         6           3103100.              0.0000120       19.6 
#>  2 Q8TBM8    0.333     6           3374207.              0.0000131        8.72
#>  3 Q5JUW8    0         6           1745373.              0.00000676       7.98
#>  4 O95833    0.167     6           3424853.              0.0000133        7.13
#>  5 Q9Y385    0         6           1415381.              0.00000549       7.10
#>  6 Q14534    0.333     6           4560342.              0.0000177        7.02
#>  7 O00214    0.333     6           1332337.              0.00000516       6.49
#>  8 Q9Y4C2    0         6           7289389.              0.0000283        5.78
#>  9 Q8WU67    0.333     6            960150.              0.00000372       5.77
#> 10 Q9H6H4    0         6           2328479.              0.00000902       5.72
#> # ℹ 4,819 more rows
#> # ℹ 12 more variables: log2_foldchange <dbl>, p_value <dbl>, adj_p_value <dbl>,
#> #   description <chr>, gene_id_entrez <chr>, gene_id_ensemble <chr>,
#> #   gene_name <chr>, biological_process <chr>, cellular_component <chr>,
#> #   molecular_function <chr>, wiki_pathway <chr>, reactome_pathway <chr>

# Note: the userdefined function is preserved in the operations tracking
hela_proteins %>% operations()
#> ℹ Data Transformations
#>   • Data files (p97KD_HCT116_proteins.xlsx) were imported as proteins from
#>   ProteomeDiscoverer
#>   • Data reassigned sample to control where 'sample like ctl'
#>   • Data reassigned sample to knockdown where 'sample like p97'
#>   • Analysis: expression difference wilcox_test knockdown/control, p.adjust =
#>   BH

# limma expression analysis
hela_proteins %>%
   expression(knockdown/control, .method = "limma") %>%
   export_analysis(knockdown/control, .analysis = "expression")
#> # A tibble: 4,829 × 20
#>    protein imputed     n log2_foldchange foldchange average_expression
#>    <chr>     <dbl> <int>           <dbl>      <dbl>              <dbl>
#>  1 Q9NXV2    0         6            3.63      12.4            1554000.
#>  2 Q5JUW8    0         6            3.05       8.28           1039329.
#>  3 O95833    0.167     6            2.88       7.37           2160177.
#>  4 Q8TBM8    0.333     6            2.85       7.20           2144962.
#>  5 Q9Y385    0         6            2.74       6.69            938010.
#>  6 Q14534    0.333     6            2.69       6.47           3072993.
#>  7 O00214    0.333     6            2.69       6.45            896168.
#>  8 Q9Y4C2    0         6            2.53       5.77           5087951.
#>  9 P84022    0.167     6            2.43       5.40            541970.
#> 10 Q96BQ5    0.5       6            2.43       5.39            520092.
#> # ℹ 4,819 more rows
#> # ℹ 14 more variables: proportional_expression <dbl>, p_value <dbl>,
#> #   adj_p_value <dbl>, limma_t_statistic <dbl>, limma_B_statistic <dbl>,
#> #   description <chr>, gene_id_entrez <chr>, gene_id_ensemble <chr>,
#> #   gene_name <chr>, biological_process <chr>, cellular_component <chr>,
#> #   molecular_function <chr>, wiki_pathway <chr>, reactome_pathway <chr>