Write table data locally

save_table() will save a summary tibble in the root directory of the local project, based on the extension given in the file name. This is a passthrough function as it returns the original tibble.

Usage

save_table(table, file_name = NULL)

Arguments

table

a tibble

file_name

a file name with extensions one of (.csv, .tsv, .rds, .xlsx)

Value

a tibble

Examples

library(dplyr, warn.conflicts = FALSE)
library(tidyproteomics)
hela_proteins %>%
   expression(knockdown/control) %>%
   export_analysis(knockdown/control, .analysis = "expression") %>%
   save_table("expression_limma_ko_over_wt.csv")
#> ℹ Save table expression_limma_ko_over_wt.csv
#> ✔ Save table expression_limma_ko_over_wt.csv ... done
#> 
#> # A tibble: 4,829 × 18
#>    protein imputed     n avera…¹ propo…² foldc…³ log2_…⁴ p_value adj_p…⁵ descr…⁶
#>    <chr>     <dbl> <int>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl> <chr>  
#>  1 Q9NXV2    0         6  3.10e6 1.20e-5   19.6     4.30 0.00417       1 BTB/PO…
#>  2 Q8TBM8    0.333     6  3.37e6 1.31e-5    8.72    3.12 0.00437       1 DnaJ h…
#>  3 Q5JUW8    0         6  1.75e6 6.76e-6    7.98    3.00 0.0232        1 Discs,…
#>  4 O95833    0.167     6  3.42e6 1.33e-5    7.13    2.83 0.0298        1 Chlori…
#>  5 Q9Y385    0         6  1.42e6 5.49e-6    7.10    2.83 0.0126        1 Ubiqui…
#>  6 Q14534    0.333     6  4.56e6 1.77e-5    7.02    2.81 0.0154        1 Squale…
#>  7 O00214    0.333     6  1.33e6 5.16e-6    6.49    2.70 0.00818       1 Galect…
#>  8 Q9Y4C2    0         6  7.29e6 2.83e-5    5.78    2.53 0.0152        1 TRPM8 …
#>  9 Q8WU67    0.333     6  9.60e5 3.72e-6    5.77    2.53 0.0763        1 Phosph…
#> 10 Q9H6H4    0         6  2.33e6 9.02e-6    5.72    2.52 0.255         1 Recept…
#> # … with 4,819 more rows, 8 more variables: gene_id_entrez <chr>,
#> #   gene_id_ensemble <chr>, gene_name <chr>, biological_process <chr>,
#> #   cellular_component <chr>, molecular_function <chr>, wiki_pathway <chr>,
#> #   reactome_pathway <chr>, and abbreviated variable names ¹​average_expression,
#> #   ²​proportional_expression, ³​foldchange, ⁴​log2_foldchange, ⁵​adj_p_value,
#> #   ⁶​description