Summarize the data

summary() is an analysis function that computes the protein summary statistics for a given tidyproteomics data object. This is a passthrough function as it returns the original tidyproteomics data-object.

Usage

# S3 method for tidyproteomics
summary(
  data,
  by = c("global"),
  destination = c("print", "save", "return"),
  limit = 25,
  contamination = NULL
)

Arguments

data

tidyproteomics data object

by

what to summarize

destination

character string, one of (save, print)

contamination

as character string

Value

a tibble on print, a tidyproteomics data-object on save

Examples

library(dplyr, warn.conflicts = FALSE)
library(tidyproteomics)

# a global summary
hela_proteins %>% summary()
#> 
#> ── Summary: global ──
#> 
#>  files protein_groups proteins peptides peptides_unique quantifiable       CVs
#>      6           8851     8851    95877           83161         75.2 0.2493123
#> 

# a summary by sample
hela_proteins %>% summary("sample")
#> 
#> ── Summary: sample ──
#> 
#>     sample files protein_groups proteins peptides peptides_unique quantifiable
#>    control     3           8851     8851    95877           83161         75.1
#>  knockdown     3           8851     8851    95877           83161         75.3
#>        CVs
#>  0.1536197
#>  0.2120776
#> 

# a summary by sample with match_between_runs removed
hela_proteins %>%
   subset(match_between_runs == FALSE) %>%
   summary("sample")
#> 
#> ℹ Subsetting data: match_between_runs == FALSE
#> ✔ Subsetting data: match_between_runs == FALSE ... done
#> 
#> 
#> ── Summary: sample ──
#> 
#>     sample files protein_groups proteins peptides peptides_unique quantifiable
#>    control     2           8222     8222    92037           81343         78.2
#>  knockdown     2           8307     8307    92584           81592         78.1
#>        CVs
#>  0.1111203
#>  0.1034109
#> 

# a summary of match_between_runs
hela_proteins %>% summary("match_between_runs")
#> 
#> ── Summary: match_between_runs ──
#> 
#>  match_between_runs files protein_groups proteins peptides peptides_unique
#>               FALSE     4           8308     8308    92494           81445
#>                TRUE     6           8851     8851    34337           28850
#>  quantifiable       CVs
#>          77.9 0.2487952
#>          70.0 0.1512411
#> 

hela_proteins %>% summary("cellular_component")
#> 
#> ── Summary: cellular_component ──
#> 
#>     cellular_component files protein_groups proteins peptides peptides_unique
#>                  Golgi     6              8        8       40              34
#>           cell surface     6             25       25      114             112
#>             chromosome     6              4        4       15              15
#>              cytoplasm     6              1        1        1               1
#>           cytoskeleton     6            201      201      973             792
#>                cytosol     6            161      161     1875            1705
#>  endoplasmic reticulum     6             44       44      499             276
#>               endosome     6            138      138     1409            1262
#>          extracellular     6             40       40      496             412
#>               membrane     6            590      590     6331            5543
#>          mitochondrion     6             19       19      172             157
#>                nucleus     6            273      273     2431            1846
#>  quantifiable       CVs
#>          85.0 0.1880712
#>          71.3 0.2765810
#>           0.0        NA
#>           0.0        NA
#>          54.4 0.2706304
#>          79.8 0.2218398
#>          47.5 0.2565168
#>          80.3 0.2349774
#>          76.6 0.2500972
#>          75.9 0.2411769
#>          83.7 0.3825493
#>          58.8 0.2764623
#> 

hela_proteins %>% summary("biological_process")
#> 
#> ── Summary: biological_process ──
#> 
#>                biological_process files protein_groups proteins peptides
#>                cell communication     6             13       13      117
#>                        cell death     6              1        1        3
#>              cell differentiation     6              3        3       22
#>                       cell growth     6            152      152     2485
#>  cell organization and biogenesis     6             23       23      292
#>                cell proliferation     6           8851     8851    95877
#>       cellular component movement     6             10       10       72
#>              cellular homeostasis     6            397      397     4071
#>                       coagulation     6              9        9      129
#>                       conjugation     6            232      232     2133
#>                  defense response     6             21       21      184
#>                       development     6             58       58      474
#>                 metabolic process     6            481      481     4536
#>  peptides_unique quantifiable       CVs
#>              110         83.8 0.2108179
#>                1         33.3 0.3560724
#>               22         18.2 0.4976946
#>             1322         46.2 0.2708449
#>              276         88.1 0.1892153
#>            83161         75.2 0.2493123
#>               55         60.4 0.2610407
#>             3707         82.6 0.2529687
#>              108         74.9 0.1999334
#>             1750         70.8 0.2561487
#>              162         71.5 0.3469895
#>              382         59.8 0.2715943
#>             3762         68.5 0.2548926
#>