Helper function to convert the data-object into a tibble

as.data.frame() is a function that converts the tidyproteomics data object into a tibble. This tibble is in the long-format, such that a there is a single observation per line.

Usage

# S3 method for tidyproteomics
as.data.frame(data, shape = c("long", "wide"), values = NULL)

Arguments

data

tidyproteomics data object

shape

the orientation of the quantitative data as either a single measure per row (long), or as multiple measures per protein/peptide (wide).

values

indicates the selected normalization to output. The default is that selected at the time of normalization.

Value

a tibble

Examples

library(dplyr, warn.conflicts = FALSE)
library(tidyproteomics)

# convert the data-object to a data.frame
hela_proteins %>% as.data.frame() %>% as_tibble()
#> # A tibble: 42,330 × 20
#>    sample_id sample_file sample  repli…¹ protein abund…² descr…³ gene_…⁴ gene_…⁵
#>    <chr>     <chr>       <chr>   <chr>   <chr>     <dbl> <chr>   <chr>   <chr>  
#>  1 e9b20ea7  f1          control 1       Q15149   1.01e9 Plecti… 5339    ENSG00…
#>  2 e9b20ea7  f1          control 1       Q09666   6.59e8 Neurob… 79026   ENSG00…
#>  3 e9b20ea7  f1          control 1       H0YDN1   2.40e5 Plecti… NA      ENSG00…
#>  4 e9b20ea7  f1          control 1       Q14204   3.86e8 Cytopl… 1778    ENSG00…
#>  5 e9b20ea7  f1          control 1       P49327   1.44e9 Fatty … 2194    ENSG00…
#>  6 e9b20ea7  f1          control 1       P78527   6.23e8 DNA-de… 5591    ENSG00…
#>  7 e9b20ea7  f1          control 1       P21333   6.05e8 Filami… 2316    ENSG00…
#>  8 e9b20ea7  f1          control 1       P35579   1.07e9 Myosin… 4627    ENSG00…
#>  9 e9b20ea7  f1          control 1       O75369   3.26e8 Filami… 2317    ENSG00…
#> 10 e9b20ea7  f1          control 1       A0A0D9…  1.96e8 Spectr… NA      ENSG00…
#> # … with 42,320 more rows, 11 more variables: gene_name <chr>,
#> #   biological_process <chr>, cellular_component <chr>,
#> #   molecular_function <chr>, wiki_pathway <chr>, reactome_pathway <chr>,
#> #   protein_group <chr>, match_between_runs <lgl>, num_peptides <dbl>,
#> #   num_unique_peptides <dbl>, num_psms <dbl>, and abbreviated variable names
#> #   ¹​replicate, ²​abundance_raw, ³​description, ⁴​gene_id_entrez,
#> #   ⁵​gene_id_ensemble

# select the wide format
hela_proteins %>% as.data.frame(shape = 'wide') %>% as_tibble()
#> # A tibble: 9,434 × 21
#>    protein    descript…¹ gene_…² gene_…³ gene_…⁴ biolo…⁵ cellu…⁶ molec…⁷ wiki_…⁸
#>    <chr>      <chr>      <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>  
#>  1 Q15149     Plectin O… 5339    ENSG00… PLEC    cell d… cytopl… motor … Envelo…
#>  2 Q09666     Neuroblas… 79026   ENSG00… AHNAK   metabo… cell s… struct… NA     
#>  3 H0YDN1     Plectin O… NA      ENSG00… PLEC    cell g… cytoso… motor … NA     
#>  4 Q14204     Cytoplasm… 1778    ENSG00… DYNC1H1 cell d… cell s… motor … Ciliar…
#>  5 P49327     Fatty aci… 2194    ENSG00… FASN    cellul… cytopl… struct… AMP-ac…
#>  6 P78527     DNA-depen… 5591    ENSG00… PRKDC   cell d… cytoso… RNA bi… miRNA …
#>  7 P21333     Filamin-A… 2316    ENSG00… FLNA    cell d… cell s… antiox… Androg…
#>  8 P35579     Myosin-9 … 4627    ENSG00… MYH9    cell d… cytopl… antiox… Primar…
#>  9 O75369     Filamin-B… 2317    ENSG00… FLNB    cell g… cytopl… motor … Focal …
#> 10 A0A0D9SGF6 Spectrin … NA      ENSG00… SPTAN1  cell g… endopl… motor … NA     
#> # … with 9,424 more rows, 12 more variables: reactome_pathway <chr>,
#> #   protein_group <chr>, match_between_runs <lgl>, num_peptides <dbl>,
#> #   num_unique_peptides <dbl>, num_psms <dbl>, control_1_raw <dbl>,
#> #   control_2_raw <dbl>, control_3_raw <dbl>, knockdown_1_raw <dbl>,
#> #   knockdown_2_raw <dbl>, knockdown_3_raw <dbl>, and abbreviated variable
#> #   names ¹​description, ²​gene_id_entrez, ³​gene_id_ensemble, ⁴​gene_name,
#> #   ⁵​biological_process, ⁶​cellular_component, ⁷​molecular_function, …