Summarize the data

expression() is an analysis function that computes the protein summary statistics for a given tidyproteomics data object.

Usage

expression(data = NULL, ..., .method = stats::t.test, .p.adjust = "BH")

Arguments

data

tidyproteomics data object

...

two sample comparison e.g. experimental/control

.method

a two-distribution test function returning a p_value for the null hypothesis. Example functions include t.test, wilcox.test, stats::ks.test, additionally, the string "limma" can be used to select from the limma package to compute an empirical Bayesian estimation which performs better with non-linear distributions and uneven replicate balance between samples.

.p.adjust

a stats::p.adjust string for multiple test correction, default is 'BH' (Benjamini & Hochberg, 1995)

Value

a tibble

Examples

library(dplyr, warn.conflicts = FALSE)
library(tidyproteomics)

# simple t.test expression analysis
hela_proteins %>%
   expression(knockdown/control) %>%
   export_analysis(knockdown/control, .analysis = "expression")
#> # A tibble: 4,829 × 18
#>    protein imputed     n avera…¹ propo…² foldc…³ log2_…⁴ p_value adj_p…⁵ descr…⁶
#>    <chr>     <dbl> <int>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl> <chr>  
#>  1 Q9NXV2    0         6  3.10e6 1.20e-5   19.6     4.30 0.00417   0.442 BTB/PO…
#>  2 Q8TBM8    0.333     6  3.37e6 1.31e-5    8.72    3.12 0.00437   0.442 DnaJ h…
#>  3 Q5JUW8    0         6  1.75e6 6.76e-6    7.98    3.00 0.0232    0.442 Discs,…
#>  4 O95833    0.167     6  3.42e6 1.33e-5    7.13    2.83 0.0298    0.442 Chlori…
#>  5 Q9Y385    0         6  1.42e6 5.49e-6    7.10    2.83 0.0126    0.442 Ubiqui…
#>  6 Q14534    0.333     6  4.56e6 1.77e-5    7.02    2.81 0.0154    0.442 Squale…
#>  7 O00214    0.333     6  1.33e6 5.16e-6    6.49    2.70 0.00818   0.442 Galect…
#>  8 Q9Y4C2    0         6  7.29e6 2.83e-5    5.78    2.53 0.0152    0.442 TRPM8 …
#>  9 Q8WU67    0.333     6  9.60e5 3.72e-6    5.77    2.53 0.0763    0.442 Phosph…
#> 10 Q9H6H4    0         6  2.33e6 9.02e-6    5.72    2.52 0.255     0.508 Recept…
#> # … with 4,819 more rows, 8 more variables: gene_id_entrez <chr>,
#> #   gene_id_ensemble <chr>, gene_name <chr>, biological_process <chr>,
#> #   cellular_component <chr>, molecular_function <chr>, wiki_pathway <chr>,
#> #   reactome_pathway <chr>, and abbreviated variable names ¹​average_expression,
#> #   ²​proportional_expression, ³​foldchange, ⁴​log2_foldchange, ⁵​adj_p_value,
#> #   ⁶​description

# a wilcox.test expression analysis
hela_proteins %>%
   expression(knockdown/control, .method = stats::wilcox.test) %>%
   export_analysis(knockdown/control, .analysis = "expression")
#> # A tibble: 4,829 × 18
#>    protein imputed     n avera…¹ propo…² foldc…³ log2_…⁴ p_value adj_p…⁵ descr…⁶
#>    <chr>     <dbl> <int>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl> <chr>  
#>  1 Q9NXV2    0         6  3.10e6 1.20e-5   19.6     4.30     0.1   0.411 BTB/PO…
#>  2 Q8TBM8    0.333     6  3.37e6 1.31e-5    8.72    3.12     0.1   0.411 DnaJ h…
#>  3 Q5JUW8    0         6  1.75e6 6.76e-6    7.98    3.00     0.1   0.411 Discs,…
#>  4 O95833    0.167     6  3.42e6 1.33e-5    7.13    2.83     0.1   0.411 Chlori…
#>  5 Q9Y385    0         6  1.42e6 5.49e-6    7.10    2.83     0.1   0.411 Ubiqui…
#>  6 Q14534    0.333     6  4.56e6 1.77e-5    7.02    2.81     0.1   0.411 Squale…
#>  7 O00214    0.333     6  1.33e6 5.16e-6    6.49    2.70     0.1   0.411 Galect…
#>  8 Q9Y4C2    0         6  7.29e6 2.83e-5    5.78    2.53     0.1   0.411 TRPM8 …
#>  9 Q8WU67    0.333     6  9.60e5 3.72e-6    5.77    2.53     0.2   0.440 Phosph…
#> 10 Q9H6H4    0         6  2.33e6 9.02e-6    5.72    2.52     0.4   0.621 Recept…
#> # … with 4,819 more rows, 8 more variables: gene_id_entrez <chr>,
#> #   gene_id_ensemble <chr>, gene_name <chr>, biological_process <chr>,
#> #   cellular_component <chr>, molecular_function <chr>, wiki_pathway <chr>,
#> #   reactome_pathway <chr>, and abbreviated variable names ¹​average_expression,
#> #   ²​proportional_expression, ³​foldchange, ⁴​log2_foldchange, ⁵​adj_p_value,
#> #   ⁶​description

# a one-tailed wilcox.test expression analysis
wilcoxon_less <- function(x, y) {
   stats::wilcox.test(x, y, alternative = "less")
}
hela_proteins <- hela_proteins %>%
   expression(knockdown/control, .method = stats::wilcox.test)

hela_proteins %>% export_analysis(knockdown/control, .analysis = "expression")
#> # A tibble: 4,829 × 18
#>    protein imputed     n avera…¹ propo…² foldc…³ log2_…⁴ p_value adj_p…⁵ descr…⁶
#>    <chr>     <dbl> <int>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl> <chr>  
#>  1 Q9NXV2    0         6  3.10e6 1.20e-5   19.6     4.30     0.1   0.411 BTB/PO…
#>  2 Q8TBM8    0.333     6  3.37e6 1.31e-5    8.72    3.12     0.1   0.411 DnaJ h…
#>  3 Q5JUW8    0         6  1.75e6 6.76e-6    7.98    3.00     0.1   0.411 Discs,…
#>  4 O95833    0.167     6  3.42e6 1.33e-5    7.13    2.83     0.1   0.411 Chlori…
#>  5 Q9Y385    0         6  1.42e6 5.49e-6    7.10    2.83     0.1   0.411 Ubiqui…
#>  6 Q14534    0.333     6  4.56e6 1.77e-5    7.02    2.81     0.1   0.411 Squale…
#>  7 O00214    0.333     6  1.33e6 5.16e-6    6.49    2.70     0.1   0.411 Galect…
#>  8 Q9Y4C2    0         6  7.29e6 2.83e-5    5.78    2.53     0.1   0.411 TRPM8 …
#>  9 Q8WU67    0.333     6  9.60e5 3.72e-6    5.77    2.53     0.2   0.440 Phosph…
#> 10 Q9H6H4    0         6  2.33e6 9.02e-6    5.72    2.52     0.4   0.621 Recept…
#> # … with 4,819 more rows, 8 more variables: gene_id_entrez <chr>,
#> #   gene_id_ensemble <chr>, gene_name <chr>, biological_process <chr>,
#> #   cellular_component <chr>, molecular_function <chr>, wiki_pathway <chr>,
#> #   reactome_pathway <chr>, and abbreviated variable names ¹​average_expression,
#> #   ²​proportional_expression, ³​foldchange, ⁴​log2_foldchange, ⁵​adj_p_value,
#> #   ⁶​description

# Note: the userdefined function is preserved in the operations tracking
hela_proteins %>% operations()
#> ℹ Data Transformations
#>   • Data files (p97KD_HCT116_proteins.xlsx) were imported as proteins from
#>   ProteomeDiscoverer
#>   • Data reassigned sample to control where 'sample like ctl'
#>   • Data reassigned sample to knockdown where 'sample like p97'
#>   • Analysis: expression difference wilcox_test knockdown/control, p.adjust =
#>   BH

# limma expression analysis
hela_proteins %>%
   expression(knockdown/control, .method = "limma") %>%
   export_analysis(knockdown/control, .analysis = "expression")
#> # A tibble: 4,829 × 20
#>    protein imputed     n log2_…¹ foldc…² avera…³ propo…⁴ p_value adj_p…⁵ limma…⁶
#>    <chr>     <dbl> <int>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#>  1 Q9NXV2    0         6    3.63   12.4   1.55e6 6.14e-6 5.40e-5 9.32e-3    6.21
#>  2 Q5JUW8    0         6    3.05    8.28  1.04e6 4.11e-6 1.62e-5 4.78e-3    7.07
#>  3 O95833    0.167     6    2.88    7.37  2.16e6 8.54e-6 1.56e-6 1.13e-3    8.96
#>  4 Q8TBM8    0.333     6    2.85    7.20  2.14e6 8.48e-6 1.59e-5 4.78e-3    7.08
#>  5 Q9Y385    0         6    2.74    6.69  9.38e5 3.71e-6 6.80e-7 1.13e-3    9.71
#>  6 Q14534    0.333     6    2.69    6.47  3.07e6 1.21e-5 1.69e-7 8.16e-4   11.1 
#>  7 O00214    0.333     6    2.69    6.45  8.96e5 3.54e-6 1.09e-6 1.13e-3    9.28
#>  8 Q9Y4C2    0         6    2.53    5.77  5.09e6 2.01e-5 1.63e-6 1.13e-3    8.92
#>  9 P84022    0.167     6    2.43    5.40  5.42e5 2.14e-6 3.79e-5 7.05e-3    6.45
#> 10 Q96BQ5    0.5       6    2.43    5.39  5.20e5 2.06e-6 8.39e-6 3.68e-3    8.02
#> # … with 4,819 more rows, 10 more variables: limma_B_statistic <dbl>,
#> #   description <chr>, gene_id_entrez <chr>, gene_id_ensemble <chr>,
#> #   gene_name <chr>, biological_process <chr>, cellular_component <chr>,
#> #   molecular_function <chr>, wiki_pathway <chr>, reactome_pathway <chr>, and
#> #   abbreviated variable names ¹​log2_foldchange, ²​foldchange,
#> #   ³​average_expression, ⁴​proportional_expression, ⁵​adj_p_value,
#> #   ⁶​limma_t_statistic