Summarize the data

expression() is an analysis function that computes the protein summary statistics for a given tidyproteomics data object.

Usage

expression(data = NULL, ..., .method = stats::t.test)

Arguments

data

tidyproteomics data object

...

two sample comparison e.g. experimental/control

.method

a two-distribution test function returning a p_value for the null hypothesis. Example functions include t.test, wilcox.test, stats::ks.test, additionally, the string "limma" can be used to select from the limma package to compute an empirical Bayesian estimation which performs better with non-linear distributions and uneven replicate balance between samples.

Value

a tibble

Examples

library(dplyr, warn.conflicts = FALSE)
library(tidyproteomics)

# simple t.test expression analysis
ecoli_proteins %>%
   expression(ko/wt) %>%
   export_analysis(ko/wt, .analysis = "expression")
#> # A tibble: 2,575 × 18
#>    protein imputed     n avera…¹ propo…² foldc…³ log2_…⁴ p_value adj_p…⁵ descr…⁶
#>    <chr>     <dbl> <int>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl> <chr>  
#>  1 P76086    0         8  6.55e6 1.10e-5    4.79    2.26  0.182        1 Transc…
#>  2 P76241    0.5       8  2.50e5 4.21e-7    2.85    1.51  0.286        1 HTH-ty…
#>  3 P09152    0         8  7.49e6 1.26e-5    2.78    1.48  0.212        1 Respir…
#>  4 P0AEU3    0.25      8  1.71e6 2.89e-6    2.48    1.31  0.560        1 Histid…
#>  5 P10903    0.625     8  4.05e5 6.83e-7    2.36    1.24  0.347        1 Nitrat…
#>  6 P0A9I8    0         8  2.89e6 4.87e-6    2.28    1.19  0.0353       1 Nitrit…
#>  7 P0AF26    0.5       8  3.72e5 6.28e-7    2.22    1.15  0.308        1 Nitrat…
#>  8 P0AC33    0         8  1.65e8 2.78e-4    2.19    1.13  0.204        1 Fumara…
#>  9 P0ADT2    0         8  2.67e7 4.50e-5    2.14    1.10  0.165        1 UPF044…
#> 10 P31825    0         8  7.59e6 1.28e-5    2.11    1.07  0.291        1 tRNA1(…
#> # … with 2,565 more rows, 8 more variables: gene_id_entrez <chr>,
#> #   gene_id_ensemble <chr>, gene_name <chr>, biological_process <chr>,
#> #   cellular_component <chr>, molecular_function <chr>, wiki_pathway <chr>,
#> #   reactome_pathway <chr>, and abbreviated variable names ¹​average_expression,
#> #   ²​proportional_expression, ³​foldchange, ⁴​log2_foldchange, ⁵​adj_p_value,
#> #   ⁶​description
#> # ℹ Use `print(n = ...)` to see more rows, and `colnames()` to see all variable names

# a wilcox.test expression analysis
ecoli_proteins %>%
   expression(ko/wt, .method = stats::wilcox.test) %>%
   export_analysis(ko/wt, .analysis = "expression")
#> # A tibble: 2,575 × 18
#>    protein imputed     n avera…¹ propo…² foldc…³ log2_…⁴ p_value adj_p…⁵ descr…⁶
#>    <chr>     <dbl> <int>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl> <chr>  
#>  1 P76086    0         8  6.55e6 1.10e-5    4.79    2.26   0.2         1 Transc…
#>  2 P76241    0.5       8  2.50e5 4.21e-7    2.85    1.51   0.114       1 HTH-ty…
#>  3 P09152    0         8  7.49e6 1.26e-5    2.78    1.48   0.114       1 Respir…
#>  4 P0AEU3    0.25      8  1.71e6 2.89e-6    2.48    1.31   0.886       1 Histid…
#>  5 P10903    0.625     8  4.05e5 6.83e-7    2.36    1.24   0.4         1 Nitrat…
#>  6 P0A9I8    0         8  2.89e6 4.87e-6    2.28    1.19   0.114       1 Nitrit…
#>  7 P0AF26    0.5       8  3.72e5 6.28e-7    2.22    1.15   0.343       1 Nitrat…
#>  8 P0AC33    0         8  1.65e8 2.78e-4    2.19    1.13   0.686       1 Fumara…
#>  9 P0ADT2    0         8  2.67e7 4.50e-5    2.14    1.10   0.2         1 UPF044…
#> 10 P31825    0         8  7.59e6 1.28e-5    2.11    1.07   0.2         1 tRNA1(…
#> # … with 2,565 more rows, 8 more variables: gene_id_entrez <chr>,
#> #   gene_id_ensemble <chr>, gene_name <chr>, biological_process <chr>,
#> #   cellular_component <chr>, molecular_function <chr>, wiki_pathway <chr>,
#> #   reactome_pathway <chr>, and abbreviated variable names ¹​average_expression,
#> #   ²​proportional_expression, ³​foldchange, ⁴​log2_foldchange, ⁵​adj_p_value,
#> #   ⁶​description
#> # ℹ Use `print(n = ...)` to see more rows, and `colnames()` to see all variable names

# a one-tailed wilcox.test expression analysis
wilcoxon_less <- function(x, y) {
   stats::wilcox.test(x, y, alternative = 'less')
}
ecoli_proteins <- ecoli_proteins %>%
   expression(ko/wt, .method = stats::wilcox.test)

ecoli_proteins %>% export_analysis(ko/wt, .analysis = "expression")
#> # A tibble: 2,575 × 18
#>    protein imputed     n avera…¹ propo…² foldc…³ log2_…⁴ p_value adj_p…⁵ descr…⁶
#>    <chr>     <dbl> <int>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl> <chr>  
#>  1 P76086    0         8  6.55e6 1.10e-5    4.79    2.26   0.2         1 Transc…
#>  2 P76241    0.5       8  2.50e5 4.21e-7    2.85    1.51   0.114       1 HTH-ty…
#>  3 P09152    0         8  7.49e6 1.26e-5    2.78    1.48   0.114       1 Respir…
#>  4 P0AEU3    0.25      8  1.71e6 2.89e-6    2.48    1.31   0.886       1 Histid…
#>  5 P10903    0.625     8  4.05e5 6.83e-7    2.36    1.24   0.4         1 Nitrat…
#>  6 P0A9I8    0         8  2.89e6 4.87e-6    2.28    1.19   0.114       1 Nitrit…
#>  7 P0AF26    0.5       8  3.72e5 6.28e-7    2.22    1.15   0.343       1 Nitrat…
#>  8 P0AC33    0         8  1.65e8 2.78e-4    2.19    1.13   0.686       1 Fumara…
#>  9 P0ADT2    0         8  2.67e7 4.50e-5    2.14    1.10   0.2         1 UPF044…
#> 10 P31825    0         8  7.59e6 1.28e-5    2.11    1.07   0.2         1 tRNA1(…
#> # … with 2,565 more rows, 8 more variables: gene_id_entrez <chr>,
#> #   gene_id_ensemble <chr>, gene_name <chr>, biological_process <chr>,
#> #   cellular_component <chr>, molecular_function <chr>, wiki_pathway <chr>,
#> #   reactome_pathway <chr>, and abbreviated variable names ¹​average_expression,
#> #   ²​proportional_expression, ³​foldchange, ⁴​log2_foldchange, ⁵​adj_p_value,
#> #   ⁶​description
#> # ℹ Use `print(n = ...)` to see more rows, and `colnames()` to see all variable names

# Note: the userdefined function is preserved in the operations tracking
ecoli_proteins %>% operations()
#> ℹ Data Transformations
#>   • Data files (pd_ecoli-hint_proteins.xlsx) were imported as proteins from
#>   ProteomeDiscoverer
#>   • Data reassigned sample to ko where 'sample like hint'
#>   • Analysis: expression difference wilcox_test ko/wt

# limma expression analysis
ecoli_proteins %>%
   expression(ko/wt, .method = "limma") %>%
   export_analysis(ko/wt, .analysis = "expression")
#> # A tibble: 2,575 × 20
#>    protein imputed     n log2_…¹ foldc…² avera…³ propo…⁴ p_value adj_p…⁵ limma…⁶
#>    <chr>     <dbl> <int>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#>  1 P76086    0         8    1.67    3.19  3.54e6 6.04e-6 0.0688    0.995    2.07
#>  2 P76241    0.5       8    1.51    2.85  1.60e5 2.72e-7 0.0458    0.995    2.32
#>  3 P10903    0.625     8    1.37    2.58  2.86e5 4.88e-7 0.0910    0.995    1.96
#>  4 P09152    0         8    1.35    2.56  5.49e6 9.38e-6 0.0438    0.995    2.34
#>  5 P0A9I8    0         8    1.27    2.41  2.49e6 4.25e-6 0.00470   0.995    3.73
#>  6 P31825    0         8    1.27    2.41  4.98e6 8.50e-6 0.0810    0.995    1.96
#>  7 P33025    0         8    1.19    2.28  2.43e6 4.15e-6 0.0216    0.995    2.77
#>  8 P0ADT2    0         8    1.14    2.21  2.11e7 3.60e-5 0.0558    0.995    2.19
#>  9 P0AF26    0.5       8    1.10    2.14  2.70e5 4.61e-7 0.128     0.995    1.68
#> 10 P0AEU3    0.25      8    1.02    2.03  1.33e6 2.27e-6 0.189     0.995    1.42
#> # … with 2,565 more rows, 10 more variables: limma_B_statistic <dbl>,
#> #   description <chr>, gene_id_entrez <chr>, gene_id_ensemble <chr>,
#> #   gene_name <chr>, biological_process <chr>, cellular_component <chr>,
#> #   molecular_function <chr>, wiki_pathway <chr>, reactome_pathway <chr>, and
#> #   abbreviated variable names ¹​log2_foldchange, ²​foldchange,
#> #   ³​average_expression, ⁴​proportional_expression, ⁵​adj_p_value,
#> #   ⁶​limma_t_statistic
#> # ℹ Use `print(n = ...)` to see more rows, and `colnames()` to see all variable names