Summarize the data

summary() is an analysis function that computes the protein summary statistics for a given tidyproteomics data object. This is a passthrough function as it returns the original tidyproteomics data-object.

Usage

# S3 method for tidyproteomics
summary(object, ...)

Arguments

object

tidyproteomics data object

...

passthrough arguments

Value

a tibble on print, a tidyproteomics data-object on save

Examples

library(dplyr, warn.conflicts = FALSE)
library(tidyproteomics)

# a global summary
hela_proteins %>% summary()
#> 
#> ── Summary: global ──
#> 
#>  proteins peptides peptides_unique quantifiable  CVs
#>      7055    66329           58706        0.908 0.25
#> 

# a summary by sample
hela_proteins %>% summary("sample")
#> 
#> ── Summary: sample ──
#> 
#>     sample proteins peptides peptides_unique quantifiable  CVs
#>    control     7055    66329           58706        0.908 0.16
#>  knockdown     7055    66329           58706        0.909 0.21
#> 

# a summary by sample with imputations removed
hela_proteins %>%
   subset(imputed == 0) %>%
   summary("sample")
#> 
#> ℹ Subsetting data: imputed == 0
#> ✔ Subsetting data: imputed == 0 ... done
#> 
#> 
#> ── Summary: sample ──
#> 
#>     sample proteins peptides peptides_unique quantifiable  CVs
#>    control     7002    65418           58151        0.907 0.16
#>  knockdown     6995    65588           58254        0.908 0.21
#> 

# a summary of imputation
hela_proteins %>% summary("imputed")
#> 
#> ── Summary: imputed ──
#> 
#>  imputed proteins peptides peptides_unique quantifiable  CVs
#>        1      695      827             504        0.999 0.24
#>        0     7055    65503           58203        0.907 0.25
#> 

hela_proteins %>% summary("cellular_component")
#> 
#> ── Summary: cellular_component ──
#> 
#>     cellular_component proteins peptides peptides_unique quantifiable   CVs
#>                  Golgi        9       44              35        0.902 0.200
#>           cell surface       18       52              52        0.712 0.275
#>             chromosome        3        6               6        0.111 0.320
#>              cytoplasm     7055    66329           58706        0.908 0.250
#>           cytoskeleton      127      504             422        0.691 0.280
#>                cytosol      121     1275            1208        0.935 0.240
#>  endoplasmic reticulum       24      182              75        0.776 0.220
#>               endosome      111      975             883        0.907 0.240
#>          extracellular       33      341             296        0.819 0.310
#>               membrane      431     4082            3535        0.890 0.250
#>          mitochondrion       16       88              87        0.831 0.310
#>                nucleus      186     1169             959        0.843 0.280
#> 

hela_proteins %>% summary("biological_process")
#> 
#> ── Summary: biological_process ──
#> 
#>                biological_process proteins peptides peptides_unique
#>                cell communication        9      100              93
#>                        cell death        1        3               1
#>              cell differentiation        3        9               9
#>                       cell growth      104     1419             839
#>  cell organization and biogenesis       17      241             241
#>                cell proliferation     7055    66329           58706
#>       cellular component movement        6       13              11
#>              cellular homeostasis      324     2854            2631
#>                       coagulation        9       68              58
#>                       conjugation      181     1460            1240
#>                  defense response       15       83              76
#>                       development       38      180             164
#>                 metabolic process      342     2804            2422
#>  quantifiable   CVs
#>         0.920 0.200
#>         1.000 0.340
#>         0.389 0.305
#>         0.803 0.280
#>         0.967 0.210
#>         0.908 0.250
#>         0.679 0.410
#>         0.938 0.260
#>         0.885 0.220
#>         0.893 0.260
#>         0.709 0.275
#>         0.730 0.245
#>         0.886 0.250
#>