Summarize the data

summary() is an analysis function that computes the protein summary statistics for a given tidyproteomics data object. This is a passthrough function as it returns the original tidyproteomics data-object.

Usage

# S3 method for tidyproteomics
summary(object, ...)

Arguments

object

tidyproteomics data object

...

passthrough arguments

Value

a tibble on print, a tidyproteomics data-object on save

Examples

library(dplyr, warn.conflicts = FALSE)
library(tidyproteomics)

# a global summary
hela_proteins %>% summary()
#> 
#> ── Summary: global ──
#> 
#>  files protein_groups proteins peptides peptides_unique quantifiable       CVs
#>      6           7055     7055    66329           58706         90.8 0.2495823
#> 

# a summary by sample
hela_proteins %>% summary("sample")
#> 
#> ── Summary: sample ──
#> 
#>     sample files protein_groups proteins peptides peptides_unique quantifiable
#>    control     3           7055     7055    66329           58706         90.8
#>  knockdown     3           7055     7055    66329           58706         90.9
#>        CVs
#>  0.1600474
#>  0.2099227
#> 

# a summary by sample with match_between_runs removed
hela_proteins %>%
   subset(match_between_runs == FALSE) %>%
   summary("sample")
#> 
#> ℹ Subsetting data: match_between_runs == FALSE
#> ✔ Subsetting data: match_between_runs == FALSE ... done
#> 
#> 
#> ── Summary: sample ──
#> 
#>     sample files protein_groups proteins peptides peptides_unique quantifiable
#>    control     3           6527     6527    62871           56511         94.0
#>  knockdown     3           6556     6556    63319           56819         93.7
#>        CVs
#>  0.1566694
#>  0.2085010
#> 

# a summary of match_between_runs
hela_proteins %>% summary("match_between_runs")
#> 
#> ── Summary: match_between_runs ──
#> 
#>  match_between_runs files protein_groups proteins peptides peptides_unique
#>               FALSE     6           7055     7055    63095           56665
#>                TRUE     6           2379     2379     3235            2042
#>  quantifiable       CVs
#>          93.9 0.2469399
#>          31.5 0.2450583
#> 

hela_proteins %>% summary("cellular_component")
#> 
#> ── Summary: cellular_component ──
#> 
#>     cellular_component files protein_groups proteins peptides peptides_unique
#>                  Golgi     6              9        9       44              35
#>           cell surface     6             18       18       52              52
#>             chromosome     6              3        3        6               6
#>              cytoplasm     6           7055     7055    66329           58706
#>           cytoskeleton     6            127      127      504             422
#>                cytosol     6            121      121     1275            1208
#>  endoplasmic reticulum     6             24       24      182              75
#>               endosome     6            111      111      975             883
#>          extracellular     6             33       33      341             296
#>               membrane     6            431      431     4082            3535
#>          mitochondrion     6             16       16       88              87
#>                nucleus     6            186      186     1169             959
#>  quantifiable       CVs
#>          90.2 0.2023808
#>          71.2 0.2781153
#>          11.1 0.3231050
#>          90.8 0.2495823
#>          69.1 0.2843047
#>          93.5 0.2357274
#>          77.6 0.2180277
#>          90.7 0.2356094
#>          81.9 0.3113072
#>          89.0 0.2490861
#>          83.1 0.3134911
#>          84.3 0.2762951
#> 

hela_proteins %>% summary("biological_process")
#> 
#> ── Summary: biological_process ──
#> 
#>                biological_process files protein_groups proteins peptides
#>                cell communication     6              9        9      100
#>                        cell death     6              1        1        3
#>              cell differentiation     6              3        3        9
#>                       cell growth     6            104      104     1419
#>  cell organization and biogenesis     6             17       17      241
#>                cell proliferation     6           7055     7055    66329
#>       cellular component movement     6              6        6       13
#>              cellular homeostasis     6            324      324     2854
#>                       coagulation     6              9        9       68
#>                       conjugation     6            181      181     1460
#>                  defense response     6             15       15       83
#>                       development     6             38       38      180
#>                 metabolic process     6            342      342     2804
#>  peptides_unique quantifiable       CVs
#>               93         92.0 0.1965058
#>                1        100.0 0.3350875
#>                9         38.9 0.3040841
#>              839         80.3 0.2761302
#>              241         96.7 0.2124385
#>            58706         90.8 0.2495823
#>               11         67.9 0.4107544
#>             2631         93.8 0.2574754
#>               58         88.5 0.2210740
#>             1240         89.3 0.2629965
#>               76         70.9 0.2713523
#>              164         73.0 0.2483315
#>             2422         88.6 0.2532647
#>